refactor: 도메인 폴더 생성(배송비 관련 유틸 상수, `PriceContainer`)
