+++ src/domains/components/PriceContainer/PriceContainer.styled.ts
