+++ src/pages/cart/CartContent/OrderContent/OrderContent.tsx
@@ -1,11 +1,12 @@
 import { useOrderList } from "../../hooks/useOrderList";
 import AllCheckBox from "@/shared/components/AllCheckBox/AllCheckBox";
 import CartList from "./CartList/CartList";
-import PriceContainer from "./PriceContainer/PriceContainer";
+import PriceContainer from "@/domains/components/PriceContainer/PriceContainer";
 import OrderConfirmButton from "./OrderConfirmButton/OrderConfirmButton";
 import * as S from "./OrderContent.styled";
 import { useCartItemContext } from "../../contexts/CartItemProvider";
 import CartItemWithContext from "./CartList/CartItemWithContext/CartItemWithContext";
+import { getDeliveryPrice } from "@/domains/utils/getDeliveryPrice";
 
 export default function OrderContent() {
   const { cartItems } = useCartItemContext();
@@ -28,6 +29,17 @@ export default function OrderContent() {
     addSelectedItem(id);
   };
 
+  const deliveryPrice = getDeliveryPrice(orderTotalPrice);
+  const priceList = [
+    {
+      title: "총 주문 금액",
+      price: orderTotalPrice,
+    },
+    {
+      title: "배송비",
+      price: deliveryPrice,
+    },
+  ];
   return (
     <>
       <AllCheckBox
@@ -46,7 +58,10 @@ export default function OrderContent() {
             />
           ))}
         </CartList>
-        <PriceContainer orderTotalPrice={orderTotalPrice} />
+        <PriceContainer
+          priceList={priceList}
+          paymentPrice={orderTotalPrice + deliveryPrice}
+        />
       </S.ScrollContainer>
       <OrderConfirmButton
         orderList={orderList}
